+++ cs-mail/src/main/java/io/cloudslang/content/mail/constants/CipherSuites.java
@@ -15,4 +15,8 @@ public final class CipherSuites {
     public static final String TLS_RSA_WITH_AES_256_GCM_SHA384 = "TLS_RSA_WITH_AES_256_GCM_SHA384";
     public static final String TLS_RSA_WITH_AES_256_CBC_SHA256 = "TLS_RSA_WITH_AES_256_CBC_SHA256";
     public static final String TLS_RSA_WITH_AES_128_CBC_SHA256 = "TLS_RSA_WITH_AES_128_CBC_SHA256";
+
+    public static boolean validate(String cipherSuite) throws IllegalAccessException {
+       return true;
+    }
 }
