+++ cs-mail/src/main/java/io/cloudslang/content/mail/sslconfig/SSLUtils.java
@@ -14,22 +14,30 @@
  */
 
 
-
 package io.cloudslang.content.mail.sslconfig;
 
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.TrustManagerFactory;
-import javax.net.ssl.X509TrustManager;
+import io.cloudslang.content.mail.constants.Constants;
+import io.cloudslang.content.mail.constants.PropNames;
+import io.cloudslang.content.mail.constants.SecurityConstants;
+import io.cloudslang.content.mail.constants.TlsVersions;
+import io.cloudslang.content.mail.entities.GetMailInput;
+import io.cloudslang.content.mail.entities.MailInput;
+import io.cloudslang.content.mail.entities.SimpleAuthenticator;
+import io.cloudslang.content.mail.utils.ProxyUtils;
+import org.apache.commons.lang3.StringUtils;
+
+import javax.mail.*;
+import javax.mail.NoSuchProviderException;
+import javax.net.ssl.*;
+import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.URL;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.security.UnrecoverableKeyException;
+import java.security.*;
 import java.security.cert.CertificateException;
+import java.util.Properties;
+
+import static org.apache.commons.lang3.StringUtils.isEmpty;
 
 /**
  * Created by giloan on 11/5/2014.
@@ -81,5 +89,186 @@ public class SSLUtils {
             }
         }
         return trustmanagers;
+    }
+
+    public static Store createMessageStore(GetMailInput input) throws Exception {
+        Properties props = new Properties();
+        if (input.getTimeout() > 0) {
+            props.put(PropNames.MAIL + input.getProtocol() + PropNames.TIMEOUT, input.getTimeout());
+            ProxyUtils.setPropertiesProxy(props, input);
+        }
+        Authenticator auth = new SimpleAuthenticator(input.getUsername(), input.getPassword());
+        Store store;
+        if (input.isEnableTLS() || input.isEnableSSL()) {
+            addSSLSettings(input.isTrustAllRoots(), input.getKeystore(), input.getKeystorePassword(),
+                    StringUtils.EMPTY, StringUtils.EMPTY);
+        }
+        if (input.isEnableTLS()) {
+            store = tryTLSOtherwiseTrySSL(auth, props, input);
+        } else if (input.isEnableSSL()) {
+            store = connectUsingSSL(props, auth, input);
+        } else {
+            store = configureStoreWithoutSSL(props, auth, input);
+            store.connect();
+        }
+
+        return store;
+    }
+
+
+    public static void addSSLSettings(boolean trustAllRoots, String keystore,
+                                      String keystorePassword, String trustKeystore, String trustPassword) throws Exception {
+        boolean useClientCert = false;
+        boolean useTrustCert = false;
+
+        String separator = System.getProperty(PropNames.FILE_SEPARATOR);
+        String javaKeystore = System.getProperty(PropNames.JAVA_HOME) + separator + "lib" + separator + "security" + separator + "cacerts";
+
+        if (keystore.length() == 0 && !trustAllRoots) {
+            boolean storeExists = new File(javaKeystore).exists();
+            keystore = (storeExists) ? Constants.FILE + javaKeystore : null;
+            if (null != keystorePassword) {
+                if (StringUtils.EMPTY.equals(keystorePassword)) {
+                    keystorePassword = SecurityConstants.DEFAULT_PASSWORD_FOR_STORE;
+                }
+            }
+            useClientCert = storeExists;
+        } else {
+            if (!trustAllRoots) {
+                if (!keystore.startsWith(Constants.HTTP))
+                    keystore = Constants.FILE + keystore;
+                useClientCert = true;
+            }
+        }
+
+        if (trustKeystore.length() == 0 && !trustAllRoots) {
+            boolean storeExists = new File(javaKeystore).exists();
+            trustKeystore = (storeExists) ? Constants.FILE + javaKeystore : null;
+            if (storeExists) {
+                if (isEmpty(trustPassword)) {
+                    trustPassword = SecurityConstants.DEFAULT_PASSWORD_FOR_STORE;
+                }
+            } else {
+                trustPassword = null;
+            }
+            useTrustCert = storeExists;
+        } else {
+            if (!trustAllRoots) {
+                if (!trustKeystore.startsWith(Constants.HTTP))
+                    trustKeystore = Constants.FILE + trustKeystore;
+                useTrustCert = true;
+            }
+
+        }
+
+        TrustManager[] trustManagers = null;
+        KeyManager[] keyManagers = null;
+
+        if (trustAllRoots) {
+            trustManagers = new TrustManager[]{new EasyX509TrustManager()};
+        }
+
+        if (useTrustCert) {
+            KeyStore trustKeyStore = SSLUtils.createKeyStore(new URL(trustKeystore), trustPassword);
+            trustManagers = SSLUtils.createAuthTrustManagers(trustKeyStore);
+        }
+        if (useClientCert) {
+            KeyStore clientKeyStore = SSLUtils.createKeyStore(new URL(keystore), keystorePassword);
+            keyManagers = SSLUtils.createKeyManagers(clientKeyStore, keystorePassword);
+        }
+
+        SSLContext context = SSLContext.getInstance("TLS");
+        context.init(keyManagers, trustManagers, new SecureRandom());
+        SSLContext.setDefault(context);
+    }
+
+
+    static Store connectUsingSSL(Properties props, Authenticator auth, GetMailInput input) throws MessagingException {
+        Store store = configureStoreWithSSL(props, auth, input);
+        store.connect();
+        return store;
+    }
+
+
+    public static Store configureStoreWithSSL(Properties props, Authenticator auth, GetMailInput input) throws NoSuchProviderException {
+        props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.SOCKET_FACTORY_CLASS, SecurityConstants.SSL_SOCKET_FACTORY);
+        props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.SOCKET_FACTORY_FALLBACK, String.valueOf(false));
+        props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.PORT, String.valueOf(input.getPort()));
+        props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.SOCKET_FACTORY_PORT, String.valueOf(input.getPort()));
+        URLName url = new URLName(input.getProtocol(), input.getHostname(), input.getPort(), StringUtils.EMPTY,
+                input.getUsername(), input.getPassword());
+        Session session = Session.getInstance(props, auth);
+        return session.getStore(url);
+    }
+
+
+    public static Store tryTLSOtherwiseTrySSL(Authenticator auth, Properties props, GetMailInput input) throws MessagingException {
+        Store store = configureStoreWithTLS(props, auth, input);
+        try {
+            store.connect(input.getHostname(), input.getUsername(), input.getPassword());
+        } catch (Exception e) {
+            if (input.isEnableSSL()) {
+                clearTLSProperties(props, input);
+                store = connectUsingSSL(props, auth, input);
+            } else {
+                throw e;
+            }
+        }
+        return store;
+    }
+
+
+    public static Store configureStoreWithTLS(Properties props, Authenticator auth, MailInput input) throws NoSuchProviderException {
+        configureWithTLS(props, input);
+        Session session = Session.getInstance(props, auth);
+        return session.getStore(input.getProtocol() + SecurityConstants.SECURE_SUFFIX);
+    }
+
+
+    public static void configureWithTLS(Properties props, final MailInput input) {
+        final boolean sslEnable = false;
+        final boolean startTlsEnable = true;
+        final boolean startTlsRequired = true;
+        final SSLSocketFactory socketFactory = new TLSSocketFactory(input.getTlsVersions(), input.getAllowedCiphers());
+
+        if (!input.getTlsVersions().isEmpty()) {
+            props.setProperty(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX +
+                    PropNames.SSL_ENABLE, String.valueOf(sslEnable));
+
+            props.setProperty(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX +
+                    PropNames.START_TLS_ENABLE, String.valueOf(startTlsEnable));
+
+            props.setProperty(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX +
+                    PropNames.START_TLS_REQUIRED, String.valueOf(startTlsRequired));
+
+            props.put(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX +
+                            PropNames.SOCKET_FACTORY, socketFactory);
+        } else {
+            props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.SSL_ENABLE, String.valueOf(sslEnable));
+
+            props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.START_TLS_ENABLE, String.valueOf(startTlsEnable));
+
+            props.setProperty(PropNames.MAIL + input.getProtocol() + PropNames.START_TLS_REQUIRED, String.valueOf(startTlsRequired));
+        }
+    }
+
+
+    private static void clearTLSProperties(Properties props, GetMailInput input) {
+        props.remove(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX + PropNames.SSL_ENABLE);
+        props.remove(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX + PropNames.START_TLS_ENABLE);
+        props.remove(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX + PropNames.START_TLS_REQUIRED);
+        props.remove(PropNames.MAIL + input.getProtocol() + SecurityConstants.SECURE_SUFFIX + PropNames.SOCKET_FACTORY);
+
+        props.remove(PropNames.MAIL + input.getProtocol() + PropNames.SSL_ENABLE);
+        props.remove(PropNames.MAIL + input.getProtocol() + PropNames.START_TLS_ENABLE);
+        props.remove(PropNames.MAIL + input.getProtocol() + PropNames.START_TLS_REQUIRED);
+    }
+
+
+    static Store configureStoreWithoutSSL(Properties props, Authenticator auth, GetMailInput input) throws NoSuchProviderException {
+        props.put(PropNames.MAIL + input.getProtocol() + PropNames.HOST, input.getHostname());
+        props.put(PropNames.MAIL + input.getProtocol() + PropNames.PORT, input.getPort());
+        Session session = Session.getInstance(props, auth);
+        return session.getStore(input.getProtocol());
     }
 }
